requirements.txt
FastAPI
uvicorn
sqlalchemy
pydantic
python-dotenv
bcrypt
passlib
email-validator
pytest
httpx